# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Multi-market investment portfolio tracker API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /watchlists:
    get:
      operationId: listWatchlists
      tags: [watchlists]
      summary: List all watchlists
      responses:
        "200":
          description: Watchlists
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Watchlist"
    post:
      operationId: createWatchlist
      tags: [watchlists]
      summary: Create a watchlist
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WatchlistInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Watchlist"
  /watchlists/{watchlistId}:
    patch:
      operationId: updateWatchlist
      tags: [watchlists]
      summary: Rename a watchlist
      parameters:
        - name: watchlistId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WatchlistUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Watchlist"
    delete:
      operationId: deleteWatchlist
      tags: [watchlists]
      summary: Delete a watchlist
      parameters:
        - name: watchlistId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /watchlists/{watchlistId}/items:
    get:
      operationId: listWatchlistItems
      tags: [watchlists]
      summary: List items in a watchlist
      parameters:
        - name: watchlistId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/WatchlistItem"
    post:
      operationId: addWatchlistItem
      tags: [watchlists]
      summary: Add a symbol to a watchlist
      parameters:
        - name: watchlistId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WatchlistItemInput"
      responses:
        "201":
          description: Added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WatchlistItem"
  /watchlists/{watchlistId}/items/{itemId}:
    delete:
      operationId: removeWatchlistItem
      tags: [watchlists]
      summary: Remove an item from a watchlist
      parameters:
        - name: watchlistId
          in: path
          required: true
          schema:
            type: integer
        - name: itemId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Removed

  # ── PRICE ALERTS ──────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Watchlist:
      type: object
      required: [id, userId, name, market, itemCount, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: string
        name:
          type: string
        market:
          type: string
          enum: [PSE, LSE, US, CRYPTO, MIXED]
        itemCount:
          type: integer
        createdAt:
          type: string
    WatchlistInput:
      type: object
      required: [name, market]
      properties:
        name:
          type: string
          minLength: 1
        market:
          type: string
          enum: [PSE, LSE, US, CRYPTO, MIXED]
    WatchlistUpdate:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        market:
          type: string
          enum: [PSE, LSE, US, CRYPTO, MIXED]
    WatchlistItem:
      type: object
      required: [id, watchlistId, symbol, name, market, currentPrice, createdAt]
      properties:
        id:
          type: integer
        watchlistId:
          type: integer
        symbol:
          type: string
        name:
          type: string
        market:
          type: string
          enum: [PSE, LSE, US, CRYPTO, CUSTOM]
        currentPrice:
          type: number
        priceChange:
          type: number
        priceChangePct:
          type: number
        currency:
          type: ["string", "null"]
        lastUpdated:
          type: ["string", "null"]
        createdAt:
          type: string
    WatchlistItemInput:
      type: object
      required: [symbol, name, market]
      properties:
        symbol:
          type: string
        name:
          type: string
        market:
          type: string
          enum: [PSE, LSE, US, CRYPTO, CUSTOM]
        currency:
          type: string

    # ── Price Alert ────────────────────────────────────────────────────────────
...